kernel-development: Match upstream remote names
The directions currently contain two different names for the upstream
remote:
  1. git remote add linus ...
  2. git remote add upstream ...

This means that when following the instructions, it's possible for
developers to end up with different names for the same remote or for the
directions to not work properly if names don't match correctly.

Update the |git remote add| commands to all use "upstream" so they are
consistent with each other.

BUG=None

Change-Id: Ie19383cd74dbaec82140013348c78bb6d606c765
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/5589487
Commit-Queue: James Hawkins <jhawkins@chromium.org>
Reviewed-by: James Hawkins <jhawkins@chromium.org>