Fix description of EntryHash in SimpleCache design doc
The SimpleCache design doc said that the EntryHash was a 40-bit truncated
per-user-seeded SHA-256 hash. It's actually a 64-bit truncated unseeded
SHA-1 hash. Correct it.

Other miscellaneous minor fixes, including mentioning that SimpleCache is
shipping everywhere but Windows and this design doc is out-of-date.

BUG=347026230

Change-Id: I1591a5f68c457a8ceaff248db4f5fc502aadc631
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/5633088
Commit-Queue: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>