Upgrade node to 20.11.0.
This updates the binaries used to build the website to
Node v20.11.0. This also switches how the binaries are fetched
to use the first-class support for GCS packages that has been
added to gclient.

The new versions of node (which are the same that Chromium uses)
do not contain an `npm` binary. I've added npm to the node_modules
accordingly, but this seems a little dodgy as it might result in
a version of npm that is not compatible with the node version.
Probably we should rewrite things to not rely on npm being installed
during the build, and then just make updating node modules w/ npm
be handled separately, like Chromium does.

Change-Id: Id1758acc72e8634c6ac1a932254193a024bf9b21
Reviewed-on: https://chromium-review.googlesource.com/c/website/+/5742613
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Joanna Wang <jojwang@chromium.org>